Commit CreateWorld function
Memory is allocated to hold a world with the number of rows and
columns in the settings, plus the extra padding around the edges.
`memset` is used to initially fill the world with '-'. While the
world file is read, '-' are skipped and '*' are written to the
file. This is a cleaner approach as it eliminates redundant writing
operations and decreases the chance to commit any unexpected,
unwanted errors.