Bug: Fix reading file lines into array
When using `getline` to read lines from the world file, `line` was
freed after outter loop was done. This caused some of the contents
of previous lines to persist and get written into the array.

This issue was resolved by freeing the line after completion of
every inner loop iteration.