Adding Options to Command Line
Implemented option parsing to improve the flexibility of the user when
interacting with the game. This was motivated because we were not able
to specify some options without being forced to specify all of them.

Using the <getopt.h>, both short and long options are supported by the
game. Additionally, a `debug` option has added in case it could be
useful in a future scenario.