Modify PrintWorld to Animate Output
To improve the experience of the user and allow them to better see the
evolution of the world under the game's constraints, the output is now
animated. The --debug flag is used to cancel the animation, which allows
a faster execution of the tests.

Additionally, the test.sh script now runs all tests using the --debug
flag.